# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcomVina API",
        "description": "Tài liệu API cho dự án EcomVina - Laravel + Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost server"
        }
    ],
    "paths": {
...
        "/api/sanphams": {
            "get": {
                "tags": [
                    "Sản phẩm"
                ],
                "summary": "Lấy danh sách sản phẩm (có lọc + phân trang)",
                "operationId": "ffe944329727765475d9564aa5dc722b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "thuonghieu",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trangthai",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm"
                    }
                }
            }
        },
        "/api/sanphams/{id}": {
            "get": {
                "tags": [
                    "Sản phẩm"
                ],
                "summary": "Xem chi tiết sản phẩm",
                "operationId": "7857bf29a46713152e07dbd6e1552b62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết sản phẩm"
                    }
                }
            }
        },
...
    },
...
}